beckend/requirements.txt
fastapi
uvicorn
bcrypt
pyjwt